Sépare les configurations Webpack de mon-entreprise et publicodes
