+++ mon-entreprise/webpack.dev.js
@@ -2,8 +2,12 @@
 
 const { map } = require('ramda')
 const webpack = require('webpack')
-const { commonLoaders, styleLoader } = require('../webpack/common')
-const { HTMLPlugins, default: common } = require('./webpack.common')
+const {
+	HTMLPlugins,
+	styleLoader,
+	commonLoaders,
+	default: common,
+} = require('./webpack.common')
 
 module.exports = {
 	...common,
